Add npm.js file and install nodemon
